requirements.txt
attrs==21.2.0
autopep8==1.5.7
bandit==1.7.0
click==8.0.1
comment-parser==1.2.3
Flask==2.0.1
gitdb==4.0.7
GitPython==3.1.24
graphviz==0.17
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==4.1.2
MarkupSafe==2.0.1
packaging==21.0
pbr==5.6.0
pluggy==1.0.0
py==1.10.0
pycodestyle==2.7.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
python-magic==0.4.18
PyYAML==6.0
six==1.16.0
smmap==4.0.0
stevedore==3.4.0
threatspec==0.5.0
toml==0.10.2
typing-extensions==3.10.0.2
Werkzeug==2.0.1
